@@ -1,7 +1,7 @@
 {
   "name": "async",
   "description": "Higher-order functions and common patterns for asynchronous code",
-  "version": "3.2.0",
+  "version": "3.2.4",
   "main": "dist/async.js",
   "author": "Caolan McMahon",
   "homepage": "https://caolan.github.io/async/",
@@ -18,55 +18,48 @@
     "module",
     "utility"
   ],
-  "dependencies": {},
   "devDependencies": {
+    "@babel/eslint-parser": "^7.16.5",
     "babel-core": "^6.26.3",
-    "babel-eslint": "^8.2.6",
     "babel-minify": "^0.5.0",
-    "babel-plugin-add-module-exports": "^0.2.1",
-    "babel-plugin-istanbul": "^5.1.4",
+    "babel-plugin-add-module-exports": "^1.0.4",
+    "babel-plugin-istanbul": "^6.1.1",
     "babel-plugin-syntax-async-generators": "^6.13.0",
     "babel-plugin-transform-es2015-modules-commonjs": "^6.26.2",
     "babel-preset-es2015": "^6.3.13",
     "babel-preset-es2017": "^6.22.0",
     "babel-register": "^6.26.0",
-    "babelify": "^8.0.0",
+    "babelify": "^10.0.0",
     "benchmark": "^2.1.1",
     "bluebird": "^3.4.6",
-    "browserify": "^16.2.3",
+    "browserify": "^17.0.0",
     "chai": "^4.2.0",
     "cheerio": "^0.22.0",
-    "coveralls": "^3.0.4",
-    "es6-promise": "^2.3.0",
-    "eslint": "^6.0.1",
-    "eslint-plugin-prefer-arrow": "^1.1.5",
-    "fs-extra": "^0.26.7",
+    "es6-promise": "^4.2.8",
+    "eslint": "^8.6.0",
+    "eslint-plugin-prefer-arrow": "^1.2.3",
+    "fs-extra": "^10.0.0",
     "jsdoc": "^3.6.2",
-    "karma": "^4.1.0",
-    "karma-browserify": "^5.3.0",
-    "karma-edge-launcher": "^0.4.2",
-    "karma-firefox-launcher": "^1.1.0",
-    "karma-junit-reporter": "^1.2.0",
-    "karma-mocha": "^1.2.0",
+    "karma": "^6.3.12",
+    "karma-browserify": "^8.1.0",
+    "karma-firefox-launcher": "^2.1.2",
+    "karma-mocha": "^2.0.1",
     "karma-mocha-reporter": "^2.2.0",
     "karma-safari-launcher": "^1.0.0",
     "mocha": "^6.1.4",
-    "mocha-junit-reporter": "^1.18.0",
     "native-promise-only": "^0.8.0-a",
-    "nyc": "^14.1.1",
-    "rimraf": "^2.5.0",
-    "rollup": "^0.63.4",
-    "rollup-plugin-node-resolve": "^2.0.0",
+    "nyc": "^15.1.0",
+    "rollup": "^2.66.1",
+    "rollup-plugin-node-resolve": "^5.2.0",
     "rollup-plugin-npm": "^2.0.0",
-    "rsvp": "^3.0.18",
-    "semver": "^5.5.0",
-    "yargs": "^11.0.0"
+    "rsvp": "^4.8.5",
+    "semver": "^7.3.5",
+    "yargs": "^17.3.1"
   },
   "scripts": {
     "coverage": "nyc npm run mocha-node-test -- --grep @nycinvalid --invert",
-    "coveralls": "npm run coverage && nyc report --reporter=text-lcov | coveralls",
     "jsdoc": "jsdoc -c ./support/jsdoc/jsdoc.json && node support/jsdoc/jsdoc-fix-html.js",
-    "lint": "eslint --fix lib/ test/ perf/memory.js perf/suites.js perf/benchmark.js support/build/ support/*.js karma.conf.js",
+    "lint": "eslint --fix .",
     "mocha-browser-test": "karma start",
     "mocha-node-test": "mocha",
     "mocha-test": "npm run mocha-node-test && npm run mocha-browser-test",
