@@ -42,7 +44,7 @@
     "test": "./test"
   },
   "engines": {
-    "node": ">= 10.12.0"
+    "node": ">= 14.0.0"
   },
   "scripts": {
     "prepublishOnly": "nps test clean build",
@@ -55,99 +57,88 @@
     "@ungap/promise-all-settled": "1.1.2",
     "ansi-colors": "4.1.1",
     "browser-stdout": "1.3.1",
-    "chokidar": "3.5.1",
-    "debug": "4.3.1",
+    "chokidar": "3.5.3",
+    "debug": "4.3.4",
     "diff": "5.0.0",
     "escape-string-regexp": "4.0.0",
     "find-up": "5.0.0",
-    "glob": "7.1.6",
-    "growl": "1.10.5",
+    "glob": "7.2.0",
     "he": "1.2.0",
-    "js-yaml": "4.0.0",
-    "log-symbols": "4.0.0",
-    "minimatch": "3.0.4",
+    "js-yaml": "4.1.0",
+    "log-symbols": "4.1.0",
+    "minimatch": "5.0.1",
     "ms": "2.1.3",
-    "nanoid": "3.1.20",
-    "serialize-javascript": "5.0.1",
+    "nanoid": "3.3.3",
+    "serialize-javascript": "6.0.0",
     "strip-json-comments": "3.1.1",
     "supports-color": "8.1.1",
-    "which": "2.0.2",
-    "wide-align": "1.1.3",
-    "workerpool": "6.1.0",
+    "workerpool": "6.2.1",
     "yargs": "16.2.0",
     "yargs-parser": "20.2.4",
     "yargs-unparser": "2.0.0"
   },
   "devDependencies": {
-    "@11ty/eleventy": "^0.11.0",
-    "@11ty/eleventy-plugin-inclusive-language": "^1.0.0",
-    "@babel/preset-env": "^7.12.17",
-    "@mocha/docdash": "^3.0.1",
-    "@rollup/plugin-babel": "^5.1.0",
-    "@rollup/plugin-commonjs": "^14.0.0",
+    "@11ty/eleventy": "^1.0.0",
+    "@11ty/eleventy-plugin-inclusive-language": "^1.0.3",
+    "@babel/eslint-parser": "^7.16.5",
+    "@mocha/docdash": "^4.0.1",
+    "@rollup/plugin-commonjs": "^21.0.2",
     "@rollup/plugin-json": "^4.1.0",
-    "@rollup/plugin-multi-entry": "^3.0.1",
-    "@rollup/plugin-node-resolve": "^8.4.0",
-    "assetgraph-builder": "^8.1.0",
+    "@rollup/plugin-multi-entry": "^4.0.1",
+    "@rollup/plugin-node-resolve": "^13.1.3",
+    "assetgraph-builder": "^9.0.0",
     "autoprefixer": "^9.8.6",
-    "babel-eslint": "^10.1.0",
-    "canvas": "^2.6.1",
-    "chai": "^4.2.0",
-    "coffee-script": "^1.12.7",
+    "canvas": "^2.9.0",
+    "chai": "^4.3.4",
+    "coffeescript": "^2.6.1",
     "configstore": "^5.0.1",
-    "core-js": "^3.6.5",
-    "coveralls": "^3.1.0",
+    "coveralls": "^3.1.1",
     "cross-env": "^7.0.2",
-    "cross-spawn": "^7.0.3",
-    "eslint": "^7.8.1",
-    "eslint-config-prettier": "^6.11.0",
-    "eslint-config-semistandard": "^15.0.1",
-    "eslint-config-standard": "^14.1.1",
-    "eslint-plugin-import": "^2.22.0",
-    "eslint-plugin-node": "^11.0.0",
-    "eslint-plugin-prettier": "^3.1.4",
-    "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-standard": "^4.0.1",
+    "eslint": "^7.32.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-config-semistandard": "^16.0.0",
+    "eslint-config-standard": "^16.0.3",
+    "eslint-plugin-import": "^2.24.2",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-promise": "^5.1.0",
     "fail-on-errors-webpack-plugin": "^3.0.0",
-    "fs-extra": "^9.0.1",
+    "fs-extra": "^10.0.0",
     "husky": "^4.2.5",
-    "hyperlink": "^4.5.2",
-    "jsdoc": "^3.6.5",
-    "jsdoc-ts-utils": "^1.1.2",
-    "karma": "^5.1.1",
+    "hyperlink": "^5.0.4",
+    "jsdoc": "^3.6.7",
+    "jsdoc-ts-utils": "^2.0.1",
+    "karma": "^6.3.11",
     "karma-chrome-launcher": "^3.1.0",
     "karma-mocha": "^2.0.1",
     "karma-mocha-reporter": "^2.2.5",
-    "karma-requirejs": "^1.1.0",
-    "karma-sauce-launcher": "^4.3.4",
+    "karma-sauce-launcher": "^4.3.6",
     "lint-staged": "^10.2.11",
-    "markdown-it": "^11.0.0",
-    "markdown-it-anchor": "^5.3.0",
-    "markdown-it-attrs": "^3.0.3",
-    "markdown-it-emoji": "^1.4.0",
-    "markdown-it-prism": "^2.1.1",
+    "markdown-it": "^12.3.2",
+    "markdown-it-anchor": "^8.4.1",
+    "markdown-it-attrs": "^4.1.3",
+    "markdown-it-emoji": "^2.0.0",
+    "markdown-it-prism": "^2.2.2",
     "markdown-toc": "^1.2.0",
-    "markdownlint-cli": "^0.23.2",
+    "markdownlint-cli": "^0.30.0",
     "needle": "^2.5.0",
     "nps": "^5.10.0",
     "nyc": "^15.1.0",
     "pidtree": "^0.5.0",
-    "prettier": "^1.19.1",
-    "remark": "^12.0.1",
-    "remark-github": "^9.0.1",
-    "remark-inline-links": "^4.0.0",
-    "requirejs": "^2.3.6",
+    "prettier": "^2.4.1",
+    "remark": "^14.0.2",
+    "remark-github": "^11.2.2",
+    "remark-inline-links": "^6.0.1",
     "rewiremock": "^3.14.3",
     "rimraf": "^3.0.2",
-    "rollup": "^2.23.1",
+    "rollup": "^2.70.1",
     "rollup-plugin-node-globals": "^1.4.0",
-    "rollup-plugin-node-polyfills": "^0.2.1",
-    "rollup-plugin-visualizer": "^4.1.0",
+    "rollup-plugin-polyfill-node": "^0.8.0",
+    "rollup-plugin-visualizer": "^5.6.0",
     "sinon": "^9.0.3",
     "strip-ansi": "^6.0.0",
     "svgo": "^1.3.2",
     "through2": "^4.0.2",
-    "to-vfile": "^6.1.0",
     "touch": "^3.1.0",
     "unexpected": "^11.14.0",
     "unexpected-eventemitter": "^2.2.0",
@@ -157,13 +148,12 @@
     "update-notifier": "^4.1.0",
     "uslug": "^1.0.4",
     "uuid": "^8.3.0",
-    "watchify": "^3.11.1",
-    "webpack": "^4.44.1",
-    "webpack-cli": "^3.3.12"
+    "watchify": "^4.0.0",
+    "webpack": "^5.67.0",
+    "webpack-cli": "^4.9.1"
   },
   "files": [
-    "bin/*mocha",
-    "assets/growl/*.png",
+    "bin/*mocha*",
     "lib/**/*.{js,html,json}",
     "index.js",
     "mocha.css",
